* src/action-copy.c: Use stat.st_atime and stat.mtime instead of   stat.st_atim and stat.st_mtim.
* src/action-listing.c,
  src/vfs/plugins/localfs/localfs.c: Include dirent.h before
  smartinclude.h.

* src/dir.c: fir_dirname() fits diretory to specified width.

* src/file.c: Use '...' as file truncator. fit_filename() fits file
  name to specified width.

* src/file_panel-defact.c: Changed used fields of a stat structure.
  Improved support of Japanese alphabet.

* src/smartinclude.h: Defined _XOPEN_SOURCE.

* src/util.c: wcsfit() fits string to specified width.

* src/widgets/widget-edit.c, src/widgets/widget-window.c: Improved
  support of Japanese alphabet.

+++ src/dir.h
@@ -20,8 +20,6 @@ BEGIN_HEADER
 #include "file.h"
 #include <vfs/vfs.h>
 
-#include <dirent.h>
-
 /********
  * Macro defenitions
  */
@@ -51,7 +49,7 @@ wcdircatsubdir (const wchar_t *__name, const wchar_t *__subname);
 
 /* Fit dirname to specified length */
 void
-fit_dirname (const wchar_t *__dir_name, long __len, wchar_t *__res);
+fit_dirname (const wchar_t *__dir_name, size_t __len, wchar_t *__res);
 
 /* Strip non-directory suffix from file name */
 wchar_t*
